fix:ZENKO-3178 move delete button into properties tab + fix tests Properties and AccountHead
