+++ src/react/account/__tests__/AccountHead.test.jsx
@@ -1,6 +1,5 @@
-import { HeadCenter, HeadRight, HeadTitle } from '../../ui-elements/ListLayout';
+import { HeadCenter, HeadTitle } from '../../ui-elements/ListLayout';
 import AccountHead from '../AccountHead';
-import { Button } from '@scality/core-ui';
 import React from 'react';
 import { reduxMount } from '../../utils/test';
 import router from 'react-router';
@@ -16,19 +15,6 @@ const account1 = {
 };
 
 describe('AccountHead', () => {
-    it('should render AccountHead component with no delete button if no account props', () => {
-        const accountName = 'newAccount';
-        jest.spyOn(router, 'useParams').mockReturnValue({ accountName });
-        const { component } = reduxMount(<AccountHead />);
-
-        expect(component.find(HeadCenter)).toHaveLength(1);
-        expect(component.find(HeadTitle).text()).toContain(accountName);
-        expect(component.find(HeadCenter)).toHaveLength(1);
-
-        const button = component.find(HeadRight).find(Button);
-        expect(button).toHaveLength(0);
-    });
-
     it('should render AccountHead component', () => {
         jest.spyOn(router, 'useParams').mockReturnValue({ accountName: account1.userName });
         const { component } = reduxMount(<AccountHead account={account1}/>);
@@ -36,9 +22,5 @@ describe('AccountHead', () => {
         expect(component.find(HeadCenter)).toHaveLength(1);
         expect(component.find(HeadTitle).text()).toContain(account1.userName);
         expect(component.find(HeadCenter)).toHaveLength(1);
-
-        const button = component.find(HeadRight).find(Button);
-        expect(button).toHaveLength(1);
-        expect(button.text()).toContain('Delete Account');
     });
 });
