+++ src/react/account/details/Properties.jsx
@@ -1,9 +1,16 @@
 // @noflow
 
 import Table, * as T from '../../ui-elements/TableKeyValue';
+import { closeAccountDeleteDialog, deleteAccount, openAccountDeleteDialog } from '../../actions';
+import { useDispatch, useSelector } from 'react-redux';
 import type { Account } from '../../../types/account';
+import type { AppState } from '../../../types/state';
+import { Button } from '@scality/core-ui';
+import { ButtonContainer } from '../../ui-elements/Container';
 import { Clipboard } from '../../ui-elements/Clipboard';
+import DeleteConfirmation from '../../ui-elements/DeleteConfirmation';
 import React from 'react';
+import { TableContainer } from '../../ui-elements/Table';
 import { formatDate } from '../../utils';
 
 type Props = {
@@ -11,8 +18,31 @@ type Props = {
 };
 
 function Properties({ account }: Props) {
+    const dispatch = useDispatch();
+    const showDelete = useSelector((state: AppState) => state.uiAccounts.showDelete);
+
+    const handleDeleteClick = () => {
+        dispatch(openAccountDeleteDialog());
+    };
+
+    const handleDeleteApprove = () => {
+        if (!account) {
+            return;
+        }
+        dispatch(deleteAccount(account.userName));
+    };
+
+    const handleDeleteCancel = () => {
+        dispatch(closeAccountDeleteDialog());
+    };
+
+    // TODO: Should we let the user delete accounts that still owns buckets.
     return (
-        <div>
+        <TableContainer>
+            <DeleteConfirmation show={showDelete} cancel={handleDeleteCancel} approve={handleDeleteApprove} titleText={`Are you sure you want to delete account: ${account.userName} ?`}/>
+            <ButtonContainer>
+                <Button icon={<i className="fas fa-trash" />} onClick={handleDeleteClick} size="default" variant="danger" text='Delete Account' />
+            </ButtonContainer>
             <h3>Account details</h3>
             <Table id='account-details-table'>
                 <T.Body>
@@ -46,7 +76,7 @@ function Properties({ account }: Props) {
                     </T.Row>
                 </T.Body>
             </Table>
-        </div>
+        </TableContainer>
     );
 }
